+++ core/Parsing.SourceCode.fs
@@ -8,9 +8,9 @@ open Sgml
 
 /// Creates a parser for source code files, given a list of comment parsers
 let sourceCode
-    (commentParsers: List<Settings -> OptionParser>)
+    (commentParsers: List<Settings -> OptionParser<string,string>>)
     (settings: Settings)
-    : TotalParser =
+    : TotalParser<string> =
 
     let commentParsers =
         tryMany (List.map (fun cp -> cp settings) commentParsers)
@@ -26,7 +26,7 @@ let customLine =
     Comments.lineComment
 
 /// A standard line comment parser with the given pattern
-let line : string -> Settings -> OptionParser =
+let line : string -> Settings -> OptionParser<string,string> =
     customLine Markdown.markdown
 
 /// Block comment parser that takes a custom content parser and middle line prefixes
@@ -34,7 +34,7 @@ let customBlock =
     Comments.blockComment
 
 /// A standard block comment parser with the given start and end patterns
-let block : (string * string) -> Settings -> OptionParser =
+let block : (string * string) -> Settings -> OptionParser<string,string> =
     customBlock Markdown.markdown ("", "")
 
 
@@ -44,7 +44,7 @@ let cLine =
 
 /// C-Style block comment parser (/* ... */)
 let cBlock =
-    customBlock Markdown.markdown (@"\*?", "") (@"/\*", @"\*/")
+    customBlock Markdown.markdown ("*", "") (@"/\*", @"\*/")
 
 /// Markers for javadoc
 let javadocMarkers =
@@ -52,21 +52,21 @@ let javadocMarkers =
 
 
 /// Parser for java/javascript (also used in html)
-let java : Settings -> TotalParser =
+let java : Settings -> TotalParser<string> =
     sourceCode
-        [ customBlock DocComments.javadoc ( "\\*?", " * " ) javadocMarkers
+        [ customBlock DocComments.javadoc ( "*", " * " ) javadocMarkers
           cBlock
           customLine DocComments.javadoc "//[/!]"
           cLine
         ]
 
 /// Parser for css (also used in html)
-let css : Settings -> TotalParser =
+let css : Settings -> TotalParser<string> =
     sourceCode
-        [ customBlock DocComments.javadoc ( "\\*?", " * " ) javadocMarkers
+        [ customBlock DocComments.javadoc ( "*", " * " ) javadocMarkers
           cBlock
         ]
 
 /// Parser for html
-let html : Settings -> TotalParser =
+let html : Settings -> TotalParser<string> =
     sgml java css [||]